[Client] Initial creation of the HTML welcome page
This page contains currently elements for the intial loading state: a theme
switcher, a play button and links to the privacy policy, the license(s) of
elements reused in the game and legal mentions.

It also contains error messages when JavaScript, websockets or local storage is
disabled, but they are not displayed or displayed properly yet.