[Client] Remove localStorage requirement
As we use cookies for the game and that localStorage is only used to save the
theme set on the start page, it isn't needed to use it, as the dark theme would
be set as a fallback.

In order to show a cookies requirement message if they are disabled, we need to
remove the one of the localStorage, as when cookies access is disabled,
localStorage cannot be accessed.