[Client] Remove elements to clone in JavaScript in the game page
The elements to clone are now embedded in the corresponding JavaScript file, so
there is no need for them to be kept in the HTML page.

Also remove on the culprit img element the src attribute to avoid network
requests on an unexisting file and the alt attribute, as they are appended
dynamically with JavaScript.